Pass norm on split_words
