+++ Makefile
@@ -6,7 +6,7 @@
 #    By: ogonzale <ogonzale@student.42barcel>       +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2022/05/18 10:00:13 by ogonzale          #+#    #+#              #
-#    Updated: 2022/11/10 18:41:22 by ogonzale         ###   ########.fr        #
+#    Updated: 2022/12/02 18:56:04 by ogonzale         ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
@@ -39,7 +39,7 @@ WHITE 		:= \033[0;97m
 #Sources
 
 PARSE_DIR	:= parse/
-PARSE_FILES	:= split_cmd_line ft_split_mod split_words split_words_2 \
+PARSE_FILES	:= split_cmd_line ft_split_mod split_words split_words_2 split_words_3\
 	           split_utils split_utils_2 expand_words
 
 MAND_DIR	:= mand/
